\name{bayesdyn}
\alias{bayesdyn}
\title{Bayesian Dynamic Modelling of Time-Series Data.}
\description{This package implements a series of dynamic models for analysing and predicting time-series data.}
\details{
		\tabular{ll}{
					Package: \tab bayesdyn\cr
					Type: \tab Statistical\cr
					Version: \tab v-1\cr
					Date: \tab 2015-03-27\cr
					}
		}			
\author{
	\tabular{ll}{
		\tab Juan Carlos Martinez-Ovando [aut, cre]\cr
		\tab Juan W. Zinser-Lopez [ctb]\cr
		}
	}
\references{
    - "A Quantitative Assessment of the Influence of Crime on the Economic Growth Dynamics in Mexico in the Short-Term," Martinez-Ovando, J. C. (2014).

	- "Predictive Vector Autoregression under Instability and Informative Prior Distribution," Martinez-Ovando, J. C. (2015).

	- "Commodity Terms-of-Trade and Inflation Targeting in Mexico," Martinez-Ovando, J. C. and  J. Zinser.," Martinez-Ovando, J. C. (2015).

	- "A Note on the Diffusion of Business Cycles," Guerrero-Escobar, S. and Martinez-Ovando, J. C. (2015).

	- "Structured Dynamic Factor Modeling to Measure Nested Business Cycles," Guerrero-Escobar, S. and Martinez-Ovando, J. C. (2015).
	}